finish ex 11.2
