+++ app/src/main/java/vn/asiantech/internship/database/DatabaseHelper.java
@@ -48,7 +48,7 @@ public class DatabaseHelper extends SQLiteOpenHelper {
     }
 
     @Override
-    public void onCreate(SQLiteDatabase db) {
+    public void onCreate(SQLiteDatabase db){
         db.execSQL(CREATE_TABLE_NOTES);
     }
