+++ app/src/main/java/vn/asiantech/internship/utils/Utils.java
@@ -1,0 +1,8 @@
+package vn.asiantech.internship.utils;
+
+/**
+ * Created by root on 6/21/17.
+ */
+
+public class Utils {
+}
