+++ app/src/main/res/anim/silde_out_left.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<set xmlns:android="http://schemas.android.com/apk/res/android"
+    android:duration="@android:integer/config_mediumAnimTime">
+
+    <translate
+        android:fromXDelta="0"
+        android:startOffset="25"
+        android:toXDelta="-100%p"/>
+
+</set>
