+++ app/src/main/res/anim/slide_out_top.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<set xmlns:android="http://schemas.android.com/apk/res/android"
+    android:fillAfter="true"
+    android:fillEnabled="true">
+    <translate
+        android:duration="@android:integer/config_longAnimTime"
+        android:fromYDelta="0"
+        android:toYDelta="-100%p" />
+</set>
