+++ app/src/main/res/layout/item_note.xml
@@ -1,43 +1,65 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="horizontal">
-
-    <TextView
-        android:id="@+id/tvDate"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:layout_weight="2"
-        android:gravity="center_horizontal"
-        android:padding="@dimen/paddingSomeThing"
-        tools:text="Monday 20 jul 20:06" />
+    android:layout_height="70dp">
 
     <LinearLayout
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:layout_weight="6"
-        android:orientation="vertical">
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:orientation="horizontal">
+
+        <View
+            android:id="@+id/view"
+            android:layout_width="8dp"
+            android:layout_height="match_parent"
+            android:layout_marginBottom="25dp" />
 
         <TextView
-            android:id="@+id/tvTitle"
-            style="@style/TextHeadStyle"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            tools:text="Title demo" />
+            android:id="@+id/tvDate"
+            android:layout_width="75dp"
+            android:layout_height="70dp"
+            android:gravity="center_horizontal"
+            android:padding="@dimen/paddingSomeThing"
+            tools:text="Monday 20 jul 20:06" />
 
-        <TextView
-            android:id="@+id/tvDescription"
-            android:layout_width="wrap_content"
+        <LinearLayout
+            android:layout_width="0dp"
             android:layout_height="wrap_content"
-            tools:text="Description demo" />
+            android:layout_weight="6"
+            android:orientation="vertical"
+            android:paddingEnd="@dimen/paddingSomeThing"
+            android:paddingLeft="@dimen/paddingSomeThing"
+            android:paddingRight="@dimen/paddingSomeThing"
+            android:paddingStart="@dimen/paddingSomeThing">
+
+            <TextView
+                android:id="@+id/tvTitle"
+                style="@style/TextHeadStyle"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                tools:text="Title demo" />
+
+            <TextView
+                android:id="@+id/tvDescription"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                tools:text="Description demo" />
+        </LinearLayout>
+
+        <ImageView
+            android:id="@+id/imgAvatar"
+            android:layout_width="0dp"
+            android:layout_height="70dp"
+            android:layout_weight="2"
+            android:contentDescription="@string/textview.text.description" />
     </LinearLayout>
 
-    <ImageView
-        android:id="@+id/imgAvatar"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:layout_weight="2"
-        android:contentDescription="@string/textview.text.description" />
-</LinearLayout>
+    <RelativeLayout
+        android:id="@+id/rlLine"
+        android:layout_width="match_parent"
+        android:layout_height="1dp"
+        android:layout_alignParentBottom="true"
+        android:background="@color/colorPrimaryDark"
+        android:tag="@string/app_name" />
+</RelativeLayout>
