+++ app/src/main/res/values/strings.xml
@@ -40,4 +40,8 @@
         <item>Open Library</item>
     </string-array>
     <string name="complete">Complete !</string>
+
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
+    <string name="text.no_content">No content</string>
 </resources>
